Move editorTestUtils into Editors/__tests__/
Co-locate the editor-specific test helper with the editors it tests.
Drop completed file-structure-reorg item from agr-future-improvements.md.
