+++ src/main/cliapp/src/components/Editors/autocomplete/base/__tests__/AutocompleteMultiTableEditor.test.jsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { AutocompleteMultiTableEditor } from '../AutocompleteMultiTableEditor';
-import { makeEditorOptions, emptyErrorMessagesRef } from '../../../../../tools/jest/editorTestUtils';
+import { makeEditorOptions, emptyErrorMessagesRef } from '../../../__tests__/editorTestUtils';
 import '../../../../../tools/jest/setupTests';
 
 describe('AutocompleteMultiTableEditor', () => {
