+++ src/main/cliapp/src/components/Editors/dropdown/boolean/__tests__/BooleanTableEditor.test.jsx
@@ -1,6 +1,6 @@
 import { render, fireEvent } from '@testing-library/react';
 import { BooleanTableEditor } from '../BooleanTableEditor';
-import { makeEditorOptions, emptyErrorMessagesRef } from '../../../../../tools/jest/editorTestUtils';
+import { makeEditorOptions, emptyErrorMessagesRef } from '../../../__tests__/editorTestUtils';
 import '../../../../../tools/jest/setupTests';
 
 vi.mock('../../../../../service/useControlledVocabularyService', () => ({
