+++ src/main/cliapp/src/components/Editors/dropdown/vocabulary/__tests__/ControlledVocabularyTableEditor.test.jsx
@@ -1,6 +1,6 @@
 import { render, fireEvent } from '@testing-library/react';
 import { ControlledVocabularyTableEditor } from '../ControlledVocabularyTableEditor';
-import { makeEditorOptions, emptyErrorMessagesRef } from '../../../../../tools/jest/editorTestUtils';
+import { makeEditorOptions, emptyErrorMessagesRef } from '../../../__tests__/editorTestUtils';
 import '../../../../../tools/jest/setupTests';
 
 const mockOptions = [
